test(table): keep single wrapped AppLayout skeleton demo
