--- pages/app-layout/auto-skeleton-table-nowrap.page.tsx
@@ -1,9 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-import React from 'react';
-
-import AutoSkeletonTable from './auto-skeleton-table';
-
-export default function AutoSkeletonTableNowrapPage() {
-  return <AutoSkeletonTable wrapLines={false} />;
-}
